+++ actionpack/lib/action_controller.rb
@@ -49,6 +49,7 @@ module ActionController
     autoload :Logging
     autoload :MimeResponds
     autoload :ParamsWrapper
+    autoload :CurrentTimezone
     autoload :RateLimiting
     autoload :Redirecting
     autoload :Renderers